chore: moved files to daemon package to avoid having more than one file as main package
