+++ eotsmanager/cmd/eotsd/daemon/flags.go
@@ -1,4 +1,4 @@
-package main
+package daemon
 
 const (
 	homeFlag        = "home"
